Completely break down search controller
