+++ api/catalog/api/models/audio.py
@@ -1,9 +1,9 @@
+from django.conf import settings
 from django.contrib.postgres.fields import ArrayField
 from django.db import models
 
 from uuslug import uuslug
 
-import catalog.api.controllers.search_controller as search_controller
 from catalog.api.models import OpenLedgerModel
 from catalog.api.models.media import (
     AbstractAltFile,
@@ -249,7 +249,7 @@ class MatureAudio(AbstractMatureMedia):
     """Stores all audios that have been flagged as 'mature'."""
 
     def delete(self, *args, **kwargs):
-        es = search_controller.es
+        es = settings.ES
         aud = Audio.objects.get(identifier=self.identifier)
         es_id = aud.id
         es.update(index="audio", id=es_id, body={"doc": {"mature": False}})
